# evmrs: reuse stack